+++ packages/terra-form-textarea/src/terra-dev-site/doc/example/Disabled.jsx
@@ -6,6 +6,7 @@ const DisabledExample = () => (
   <Field label="Disabled Textarea" htmlFor="disabled">
     <Textarea
       disabled
+      size="small"
       value="I'm disabled."
       id="disabled"
     />