+++ packages/terra-form-textarea/src/terra-dev-site/doc/example/FixedLargeSize.jsx
@@ -7,6 +7,7 @@ const FixedLargeSizeExample = () => (
     <Textarea
       size="large"
       id="large"
+      disableResize
     />
   </Field>
 );