+++ packages/terra-form-textarea/src/terra-dev-site/doc/example/FixedMediumSize.jsx
@@ -7,6 +7,7 @@ const FixedMediumSizeExample = () => (
     <Textarea
       size="medium"
       id="medium"
+      disableResize
     />
   </Field>
 );