+++ packages/terra-form-textarea/src/terra-dev-site/doc/example/FixedSmallSize.jsx
@@ -7,6 +7,7 @@ const FixedSmallSizeExample = () => (
     <Textarea
       size="small"
       id="small"
+      disableResize
     />
   </Field>
 );