+++ packages/terra-form-textarea/src/terra-dev-site/doc/example/Invalid.jsx
@@ -22,6 +22,7 @@ class InvalidExample extends Component {
           value={this.state.area}
           isInvalid={this.state.area === ''}
           onChange={this.handleAreaChange}
+          size="small"
           id="invalid"
         />
       </Field>