+++ packages/terra-form-textarea/src/terra-dev-site/doc/example/Resizable.jsx
@@ -6,6 +6,7 @@ const ResizableExample = () => (
   <Field label="Resizable Textarea" htmlFor="resizable">
     <Textarea
       isAutoResizable
+      size="small"
       defaultValue="I automatically resize as you type in more information (except in mobile)."
       id="resizable"
     />