+++ packages/terra-form-textarea/src/terra-dev-site/doc/example/Textarea.jsx
@@ -5,6 +5,7 @@ import Textarea from 'terra-form-textarea';
 const DefaultTextArea = () => (
   <Field label="Default Textarea">
     <Textarea
+      size="small"
       id="default"
       ariaLabel="Default Textarea"
     />